+++ tomcat/src/main/java/nextstep/handler/util/ResourceProcessor.java
@@ -1,11 +1,11 @@
-package org.apache.coyote.handler.util;
+package nextstep.handler.util;
 
 import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import org.apache.coyote.handler.util.exception.ResourceNotFoundException;
+import nextstep.handler.util.exception.ResourceNotFoundException;
 import org.apache.coyote.http.request.Request;
 import org.apache.coyote.http.response.ContentType;
 import org.apache.coyote.http.util.HttpHeaderConsts;
