+++ tomcat/src/main/java/nextstep/jwp/application/UserService.java
@@ -3,7 +3,7 @@ package nextstep.jwp.application;
 import nextstep.jwp.application.exception.AlreadyExistsAccountException;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.handler.exception.LoginFailureException;
+import nextstep.handler.exception.LoginFailureException;
 
 public class UserService {
