+++ tomcat/src/main/java/org/apache/coyote/context/HelloWorldContext.java
@@ -8,7 +8,7 @@ import org.apache.coyote.Handler;
 import org.apache.coyote.context.exception.InvalidRootContextPathException;
 import org.apache.coyote.context.exception.InvalidStaticResourcePathException;
 import org.apache.coyote.context.exception.UnsupportedApiException;
-import org.apache.coyote.handler.ResourceHandler;
+import nextstep.handler.ResourceHandler;
 import org.apache.coyote.http.SessionManager;
 import org.apache.coyote.http.request.Request;
 import org.apache.coyote.http.response.ContentType;
