+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -14,9 +14,9 @@ import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Container;
 import org.apache.coyote.Processor;
 import org.apache.coyote.context.exception.UnsupportedApiException;
-import org.apache.coyote.handler.exception.InvalidQueryParameterException;
-import org.apache.coyote.handler.exception.LoginFailureException;
-import org.apache.coyote.handler.util.exception.ResourceNotFoundException;
+import nextstep.handler.exception.InvalidQueryParameterException;
+import nextstep.handler.exception.LoginFailureException;
+import nextstep.handler.util.exception.ResourceNotFoundException;
 import org.apache.coyote.http.request.Request;
 import org.apache.coyote.http.response.ContentType;
 import org.apache.coyote.http.response.HttpStatusCode;
