+++ tomcat/src/test/java/nextstep/handler/LoginHandlerTest.java
@@ -1,7 +1,11 @@
-package org.apache.coyote.handler;
+package nextstep.handler;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.io.IOException;
+import nextstep.handler.LoginHandler;
+import nextstep.jwp.application.UserService;
+import org.apache.coyote.http.SessionManager;
 import org.apache.coyote.http.request.HttpRequestBody;
 import org.apache.coyote.http.request.HttpRequestHeaders;
 import org.apache.coyote.http.request.Parameters;
@@ -10,27 +14,26 @@ import org.apache.coyote.http.request.Url;
 import org.apache.coyote.http.response.Response;
 import org.apache.coyote.http.util.HttpMethod;
 import org.apache.coyote.http.util.HttpVersion;
-import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class WelcomeHandlerTest {
+class LoginHandlerTest {
 
     @Test
-    void 생성자는_rootContextPath를_전달하면_WelcomeHandler를_초기화한다() {
-        final WelcomeHandler actual = new WelcomeHandler();
+    void 생성자는_경로와_rootContextPath를_전달하면_LoginHandler를_초기화한다() {
+        final LoginHandler actual = new LoginHandler("/login", new UserService());
 
         assertThat(actual).isNotNull();
     }
 
     @Test
     void supports_메서드는_지원하는_요청인_경우_true를_반환한다() {
-        final WelcomeHandler handler = new WelcomeHandler();
-        final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
-        final HttpMethod method = HttpMethod.findMethod("get");
-        final Url url = Url.from("/");
+        final LoginHandler handler = new LoginHandler("/login", new UserService());
+        final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: application/json");
+        final HttpMethod method = HttpMethod.findMethod("post");
+        final Url url = Url.from("/login");
         final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
         final Request request = new Request(
                 headers,
@@ -38,20 +41,20 @@ class WelcomeHandlerTest {
                 version,
                 url,
                 HttpRequestBody.EMPTY,
-                Parameters.EMPTY
+                Parameters.fromBodyContent("account=gugu&password=password")
         );
 
-        final boolean actual = handler.supports(request,"/");
+        final boolean actual = handler.supports(request, "/");
 
         assertThat(actual).isTrue();
     }
 
     @Test
     void supports_메서드는_지원하지_않는_요청인_경우_false를_반환한다() {
-        final WelcomeHandler handler = new WelcomeHandler();
+        final LoginHandler handler = new LoginHandler("/login", new UserService());
         final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
         final HttpMethod method = HttpMethod.findMethod("get");
-        final Url url = Url.from("/");
+        final Url url = Url.from("/index.html");
         final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
         final Request request = new Request(
                 headers,
@@ -62,17 +65,17 @@ class WelcomeHandlerTest {
                 Parameters.EMPTY
         );
 
-        final boolean actual = handler.supports(request, "/hello");
+        final boolean actual = handler.supports(request, "/");
 
         assertThat(actual).isFalse();
     }
 
     @Test
-    void service_메서드는_요청을_처리하고_Response를_반환한다() {
-        final WelcomeHandler handler = new WelcomeHandler();
-        final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
-        final HttpMethod method = HttpMethod.findMethod("get");
-        final Url url = Url.from("/");
+    void service_메서드는_요청을_처리하고_Response를_반환한다() throws IOException {
+        final LoginHandler handler = new LoginHandler("/login", new UserService());
+        final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: application/json");
+        final HttpMethod method = HttpMethod.findMethod("post");
+        final Url url = Url.from("/login");
         final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
         final Request request = new Request(
                 headers,
@@ -80,19 +83,12 @@ class WelcomeHandlerTest {
                 version,
                 url,
                 HttpRequestBody.EMPTY,
-                Parameters.EMPTY
+                Parameters.fromUrlContent("?account=gugu&password=password")
         );
+        request.initSessionManager(new SessionManager());
 
         final Response actual = handler.service(request, "ignored");
 
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(actual).isNotNull();
-            softAssertions.assertThat(actual.convertResponseMessage())
-                          .contains("HTTP/1.1")
-                          .contains("200 OK")
-                          .contains("Content-Type: text/html;charset=utf-8")
-                          .contains("Content-Length: 12")
-                          .contains("Hello World!");
-        });
+        assertThat(actual.convertResponseMessage()).contains("302 Found");
     }
 }
