+++ tomcat/src/test/java/nextstep/handler/LoginPageHandlerTest.java
@@ -1,9 +1,10 @@
-package org.apache.coyote.handler;
+package nextstep.handler;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.IOException;
+import nextstep.handler.LoginPageHandler;
 import org.apache.coyote.http.SessionManager;
 import org.apache.coyote.http.request.HttpRequestBody;
 import org.apache.coyote.http.request.HttpRequestHeaders;
