+++ tomcat/src/test/java/nextstep/handler/ResourceHandlerTest.java
@@ -1,10 +1,9 @@
-package org.apache.coyote.handler;
+package nextstep.handler;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.IOException;
-import nextstep.jwp.application.UserService;
-import org.apache.coyote.http.SessionManager;
+import nextstep.handler.ResourceHandler;
 import org.apache.coyote.http.request.HttpRequestBody;
 import org.apache.coyote.http.request.HttpRequestHeaders;
 import org.apache.coyote.http.request.Parameters;
@@ -13,44 +12,24 @@ import org.apache.coyote.http.request.Url;
 import org.apache.coyote.http.response.Response;
 import org.apache.coyote.http.util.HttpMethod;
 import org.apache.coyote.http.util.HttpVersion;
+import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class RegisterHandlerTest {
+class ResourceHandlerTest {
 
     @Test
-    void 생성자는_경로와_rootContextPath를_전달하면_RegisterHandler를_초기화한다() {
-        final RegisterHandler actual = new RegisterHandler("/register", new UserService());
+    void 생성자는_static_resource의_경로를_전달하면_CssHandler를_초기화한다() {
+        final ResourceHandler actual = new ResourceHandler();
 
         assertThat(actual).isNotNull();
     }
 
     @Test
     void supports_메서드는_지원하는_요청인_경우_true를_반환한다() {
-        final RegisterHandler handler = new RegisterHandler("/register", new UserService());
-        final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: application/json");
-        final HttpMethod method = HttpMethod.findMethod("post");
-        final Url url = Url.from("/register");
-        final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
-        final Request request = new Request(
-                headers,
-                method,
-                version,
-                url,
-                HttpRequestBody.EMPTY,
-                Parameters.fromBodyContent("account=asdf&password=asdf&email=asdf@asdf.com")
-        );
-
-        final boolean actual = handler.supports(request, "/");
-
-        assertThat(actual).isTrue();
-    }
-
-    @Test
-    void supports_메서드는_지원하지_않는_요청인_경우_false를_반환한다() {
-        final RegisterHandler handler = new RegisterHandler("/register", new UserService());
+        final ResourceHandler handler = new ResourceHandler();
         final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
         final HttpMethod method = HttpMethod.findMethod("get");
         final Url url = Url.from("/index.html");
@@ -64,17 +43,17 @@ class RegisterHandlerTest {
                 Parameters.EMPTY
         );
 
-        final boolean actual = handler.supports(request, "/");
+        final boolean actual = handler.supports(request, "ignored");
 
-        assertThat(actual).isFalse();
+        assertThat(actual).isTrue();
     }
 
     @Test
-    void service_메서드는_요청을_처리하고_Response를_반환한다() throws IOException {
-        final RegisterHandler handler = new RegisterHandler("/register", new UserService());
-        final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: application/json");
-        final HttpMethod method = HttpMethod.findMethod("post");
-        final Url url = Url.from("/register");
+    void supports_메서드는_지원하지_않는_요청인_경우_false를_반환한다() {
+        final ResourceHandler handler = new ResourceHandler();
+        final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
+        final HttpMethod method = HttpMethod.findMethod("get");
+        final Url url = Url.from("/login");
         final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
         final Request request = new Request(
                 headers,
@@ -82,12 +61,35 @@ class RegisterHandlerTest {
                 version,
                 url,
                 HttpRequestBody.EMPTY,
-                Parameters.fromBodyContent("account=asdf&password=asdf&email=asdf@asdf.com")
+                Parameters.EMPTY
         );
-        request.initSessionManager(new SessionManager());
 
-        final Response actual = handler.service(request, "ignored");
+        final boolean actual = handler.supports(request, "ignored");
 
-        assertThat(actual.convertResponseMessage()).contains("302 Found");
+        assertThat(actual).isFalse();
+    }
+
+    @Test
+    void service_메서드는_요청을_처리하고_Response를_반환한다() throws IOException {
+        final ResourceHandler handler = new ResourceHandler();
+        final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
+        final HttpMethod method = HttpMethod.findMethod("get");
+        final Url url = Url.from("/index.html");
+        final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
+        final Request request = new Request(
+                headers,
+                method,
+                version,
+                url,
+                HttpRequestBody.EMPTY,
+                Parameters.EMPTY
+        );
+
+        final Response actual = handler.service(request, "static/");
+
+        SoftAssertions.assertSoftly(softAssertions -> {
+            softAssertions.assertThat(actual).isNotNull();
+            softAssertions.assertThat(actual.convertResponseMessage()).contains("<!DOCTYPE html>");
+        });
     }
 }
