+++ tomcat/src/test/java/nextstep/handler/WelcomeHandlerTest.java
@@ -1,8 +1,8 @@
-package org.apache.coyote.handler;
+package nextstep.handler;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.io.IOException;
+import nextstep.handler.WelcomeHandler;
 import org.apache.coyote.http.request.HttpRequestBody;
 import org.apache.coyote.http.request.HttpRequestHeaders;
 import org.apache.coyote.http.request.Parameters;
@@ -17,21 +17,21 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class ResourceHandlerTest {
+class WelcomeHandlerTest {
 
     @Test
-    void 생성자는_static_resource의_경로를_전달하면_CssHandler를_초기화한다() {
-        final ResourceHandler actual = new ResourceHandler();
+    void 생성자는_rootContextPath를_전달하면_WelcomeHandler를_초기화한다() {
+        final WelcomeHandler actual = new WelcomeHandler();
 
         assertThat(actual).isNotNull();
     }
 
     @Test
     void supports_메서드는_지원하는_요청인_경우_true를_반환한다() {
-        final ResourceHandler handler = new ResourceHandler();
+        final WelcomeHandler handler = new WelcomeHandler();
         final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
         final HttpMethod method = HttpMethod.findMethod("get");
-        final Url url = Url.from("/index.html");
+        final Url url = Url.from("/");
         final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
         final Request request = new Request(
                 headers,
@@ -42,17 +42,17 @@ class ResourceHandlerTest {
                 Parameters.EMPTY
         );
 
-        final boolean actual = handler.supports(request, "ignored");
+        final boolean actual = handler.supports(request,"/");
 
         assertThat(actual).isTrue();
     }
 
     @Test
     void supports_메서드는_지원하지_않는_요청인_경우_false를_반환한다() {
-        final ResourceHandler handler = new ResourceHandler();
+        final WelcomeHandler handler = new WelcomeHandler();
         final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
         final HttpMethod method = HttpMethod.findMethod("get");
-        final Url url = Url.from("/login");
+        final Url url = Url.from("/");
         final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
         final Request request = new Request(
                 headers,
@@ -63,17 +63,17 @@ class ResourceHandlerTest {
                 Parameters.EMPTY
         );
 
-        final boolean actual = handler.supports(request, "ignored");
+        final boolean actual = handler.supports(request, "/hello");
 
         assertThat(actual).isFalse();
     }
 
     @Test
-    void service_메서드는_요청을_처리하고_Response를_반환한다() throws IOException {
-        final ResourceHandler handler = new ResourceHandler();
+    void service_메서드는_요청을_처리하고_Response를_반환한다() {
+        final WelcomeHandler handler = new WelcomeHandler();
         final HttpRequestHeaders headers = HttpRequestHeaders.from("Content-Type: text/html;charset=utf-8");
         final HttpMethod method = HttpMethod.findMethod("get");
-        final Url url = Url.from("/index.html");
+        final Url url = Url.from("/");
         final HttpVersion version = HttpVersion.findVersion("HTTP/1.1");
         final Request request = new Request(
                 headers,
@@ -84,11 +84,16 @@ class ResourceHandlerTest {
                 Parameters.EMPTY
         );
 
-        final Response actual = handler.service(request, "static/");
+        final Response actual = handler.service(request, "ignored");
 
         SoftAssertions.assertSoftly(softAssertions -> {
             softAssertions.assertThat(actual).isNotNull();
-            softAssertions.assertThat(actual.convertResponseMessage()).contains("<!DOCTYPE html>");
+            softAssertions.assertThat(actual.convertResponseMessage())
+                          .contains("HTTP/1.1")
+                          .contains("200 OK")
+                          .contains("Content-Type: text/html;charset=utf-8")
+                          .contains("Content-Length: 12")
+                          .contains("Hello World!");
         });
     }
 }
